+++ src/orbitmesh-console/Scripts/app.js
@@ -22,7 +22,7 @@ const App = {
             return ["/edges", "/packages", "/variables"].includes(this.$route.path);
         },
         administrationGroupActive() {
-            return ["/repository", "/credentials", "/configuration"].includes(this.$route.path);
+            return ["/repository", "/credentials", "/configuration", "/scheduledtasks"].includes(this.$route.path);
         }
     },
     methods: {
@@ -69,6 +69,7 @@ const App = {
                             <router-link to="/repository">Repository</router-link>
                             <router-link to="/credentials">Credentials</router-link>
                             <router-link to="/configuration">Configuration</router-link>
+                            <router-link to="/scheduledtasks">Scheduled Tasks</router-link>
                         </div>
                     </div>
                 </nav>